feat: 체크리스트 배너 섹션 작업
